feat: - Rotas get e post para Salas e Reservas; - Camadas de serviço, repositórios; - Dtos, mappers; - H2, data.sql (para testes).

+++ src/main/java/com/guilhermecode/reserva_salas_api/repositories/SalaRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SalaRepository extends JpaRepository<Sala, Integer> { }
+public interface SalaRepository extends JpaRepository<Sala, Long> { }
